refactor: consolidate UI components into shadcn/ and custom/ directories
