+++ frontend/src/components/ui/custom/ChartWidget.tsx
