+++ frontend/src/components/ui/custom/LazyChartWidget.tsx
